# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface Profile {
  id: string;
  fullName: string;
  specialty?: string | null;
  crmNumber?: string | null;
  phone?: string | null;
  email?: string | null;
}

export interface ProfileUpdate {
  fullName?: string;
  specialty?: string;
  crmNumber?: string;
  phone?: string;
}

export interface AdminFlag {
  isAdmin: boolean;
}

export interface Patient {
  id: string;
  doctorId: string;
  name: string;
  age?: number | null;
  medicalRecordId?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PatientCreate {
  name: string;
  age?: number | null;
  medicalRecordId?: string;
  notes?: string;
}

export interface PatientUpdate {
  name?: string;
  age?: number | null;
  medicalRecordId?: string;
  notes?: string;
}

export interface Exam {
  id: string;
  doctorId: string;
  patientId?: string | null;
  calcType: string;
  inputData?: unknown;
  resultData?: unknown;
  gestationalAgeWeeks?: number | null;
  gestationalAgeDays?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface ExamCreate {
  patientId?: string | null;
  calcType: string;
  inputData?: unknown;
  resultData?: unknown;
  gestationalAgeWeeks?: number | null;
  gestationalAgeDays?: number | null;
  notes?: string;
}

export interface Plan {
  id: string;
  name: string;
  tier: string;
  description?: string | null;
  priceCents: number;
  durationMonths: number;
  tokensPerPeriod: number;
  features: unknown;
  isActive: boolean;
  stripePriceId?: string | null;
}

export interface SubscriptionState {
  subscribed: boolean;
  subscriptionTier?: string | null;
  subscriptionEnd?: string | null;
  tokensRemaining: number;
  isTrial: boolean;
  trialEndsAt?: string | null;
  status?: string | null;
}

export interface UseTokenResult {
  success: boolean;
  tokensRemaining: number;
}

export interface TokensRemaining {
  tokensRemaining: number;
}

export interface CheckoutRequest {
  priceId: string;
}

export interface UrlResponse {
  url: string;
}

export interface CheckoutStatusInfo {
  status: string;
  paymentStatus?: string | null;
  subscriptionId?: string | null;
}

export type AdminUserRowSubscription = {
  tier?: string | null;
  status?: string | null;
  tokensRemaining?: number;
  endDate?: string | null;
} | null;

export interface AdminUserRow {
  id: string;
  fullName?: string | null;
  email?: string | null;
  crmNumber?: string | null;
  isAdmin: boolean;
  subscription?: AdminUserRowSubscription;
}

export interface GrantTokensRequest {
  userId: string;
  amount: number;
}

export type ListExamsParams = {
patientId?: string;
};

export type CheckoutStatusParams = {
session_id: string;
};