# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: string;
  doctorId: string;
  patientId?: string | null;
  calcType: string;
  inputData?: unknown;
  resultData?: unknown;
  gestationalAgeWeeks?: number | null;
  gestationalAgeDays?: number | null;
  notes?: string | null;
  createdAt: string;
}