# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: string;
  doctorId: string;
  name: string;
  age?: number | null;
  medicalRecordId?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}