# lib/api-zod/src/generated/types/patientUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface PatientUpdate {
  name?: string;
  age?: number | null;
  medicalRecordId?: string;
  notes?: string;
}