# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  name: string;
  tier: string;
  description?: string | null;
  priceCents: number;
  durationMonths: number;
  tokensPerPeriod: number;
  features: unknown;
  isActive: boolean;
  stripePriceId?: string | null;
}