# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  fullName: string;
  specialty?: string | null;
  crmNumber?: string | null;
  phone?: string | null;
  email?: string | null;
}