# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  fullName?: string;
  specialty?: string;
  crmNumber?: string;
  phone?: string;
}