# lib/api-zod/src/generated/types/subscriptionState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionState {
  subscribed: boolean;
  subscriptionTier?: string | null;
  subscriptionEnd?: string | null;
  tokensRemaining: number;
  isTrial: boolean;
  trialEndsAt?: string | null;
  status?: string | null;
}