# lib/api-zod/src/generated/types/useTokenResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IDALIA API
 * OpenAPI spec version: 0.1.0
 */

export interface UseTokenResult {
  success: boolean;
  tokensRemaining: number;
}